requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.0.0
six==1.16.0
sqlparse==0.4.2
